+++ include/dolfin/common/system.h
@@ -11,57 +11,94 @@
 namespace dolfin
 {
 
+//-----------------------------------------------------------------------------
+
 ///
 auto filename( std::string const & name,
                std::string const & ext,
-               uint                counter = 0,
+               size_t              counter = 0,
                int                 width   = 0 ) -> std::string;
 
-auto strcounter( uint counter = 0, int width = 0 ) -> std::string;
+//-----------------------------------------------------------------------------
+
+auto strcounter( size_t counter = 0, int width = 0 ) -> std::string;
+
+//-----------------------------------------------------------------------------
 
 ///
 auto basename( std::string file ) -> std::string;
 
+//-----------------------------------------------------------------------------
+
 ///
 auto dirname( std::string file ) -> std::string;
 
+//-----------------------------------------------------------------------------
+
 ///
 auto path( std::string p0, std::string const & p1 ) -> std::string;
+
+//-----------------------------------------------------------------------------
+
 auto path( std::string p0, std::string const & p1, std::string const & p2 )
   -> std::string;
+
+//-----------------------------------------------------------------------------
 
 ///
 void glob( std::string const & pattern, std::vector< std::string > & matches );
 
+//-----------------------------------------------------------------------------
+
 ///
 void mkdir( std::string const & dir );
+
+//-----------------------------------------------------------------------------
 
 ///
 auto stat( std::string const & file ) -> bool;
 
+//-----------------------------------------------------------------------------
+
 ///
 auto getcwd() -> std::string;
+
+//-----------------------------------------------------------------------------
 
 ///
 void pwd();
 
+//-----------------------------------------------------------------------------
+
 ///
 void cd( std::string const & dir );
+
+//-----------------------------------------------------------------------------
 
 ///
 void mkdircd( std::string const & dir );
 
+//-----------------------------------------------------------------------------
+
 ///
 void pushd( std::string const & dir );
+
+//-----------------------------------------------------------------------------
 
 ///
 void popd();
 
+//-----------------------------------------------------------------------------
+
 ///
 void dirs( int n, std::string const & dir );
 
+//-----------------------------------------------------------------------------
+
 ///
 auto dirstack() -> std::vector< std::string > &;
+
+//-----------------------------------------------------------------------------
 
 ///
 template < class T >
@@ -71,7 +108,7 @@ auto human_readable( T value ) -> std::string
   static std::string const _unit[] = {
     "B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB" };
 
-  uint        i = 0;
+  size_t      i = 0;
   std::size_t x( value );
   while ( x >> 10 )
   {
@@ -82,6 +119,8 @@ auto human_readable( T value ) -> std::string
   ss << x << _unit[i];
   return ss.str();
 }
+
+//-----------------------------------------------------------------------------
 
 } /* namespace dolfin */
