+++ include/dolfin/evolution/Time.h
@@ -13,32 +13,49 @@ class Time
 {
 
 public:
-
-  using Interval = std::pair<real, real>;
+  using Interval = std::pair< real, real >;
 
   Time( real T_start = 0.0, real T_end = 0.0, real T_current = 0.0 );
 
-  Time(Interval I);
+  Time( Interval I );
 
-  Time(Time const& other) = default;
+  Time( Time const & other ) = default;
 
   ~Time() = default;
 
   /// Operators
-  inline operator real&() { return t_; }
-  inline operator real const&() const { return t_; }
-  inline auto operator=(real t) -> Time& { t_ = t; return *this; }
-  inline auto operator+=(real k) -> Time& { t_ += k; return *this; }
-  inline auto operator-=(real k) -> Time& { t_ -= k; return *this; }
+  inline operator real &()
+  {
+    return t_;
+  }
+  inline operator real const &() const
+  {
+    return t_;
+  }
+  inline auto operator=( real t ) -> Time &
+  {
+    t_ = t;
+    return *this;
+  }
+  inline auto operator+=( real k ) -> Time &
+  {
+    t_ += k;
+    return *this;
+  }
+  inline auto operator-=( real k ) -> Time &
+  {
+    t_ -= k;
+    return *this;
+  }
 
   /// Return time interval as a pair of real numbers
-  auto interval() const -> Interval const&;
+  auto interval() const -> Interval const &;
 
   /// Return time sign function
   auto sign() const -> int;
 
   /// Is the current time in the time interval
-  auto is_valid(real atol = 0.0) const -> bool;
+  auto is_valid( real atol = 0.0 ) const -> bool;
 
   /// T0
   auto begin() const -> real;
@@ -50,7 +67,7 @@ public:
   auto measure() const -> real;
 
   /// Clock
-  auto clock() const -> real const&;
+  auto clock() const -> real const &;
 
   /// Clock [Fragile]
   auto clock() -> real &;
@@ -74,7 +91,11 @@ public:
   void disp() const;
 
   /// Step
-  inline auto step(real k ) -> Time& { t_ +=  sign_ * k; return *this; }
+  inline auto step( real k ) -> Time &
+  {
+    t_ += sign_ * k;
+    return *this;
+  }
 
   //--- ITERATOR --------------------------------------------------------------
 
@@ -82,7 +103,6 @@ public:
   {
 
   public:
-
     /// Constructor
     iterator() = default;
 
@@ -92,34 +112,31 @@ public:
     //--- INTERFACE -----------------------------------------------------------
 
     /// Pre-increment
-    virtual auto operator++() -> Time& = 0;
+    virtual auto operator++() -> Time & = 0;
 
     /// Pre-decrement
-    virtual auto operator--() -> Time& = 0;
+    virtual auto operator--() -> Time & = 0;
 
     /// Is the iterator valid ?
     virtual auto is_valid() const -> bool = 0;
 
     /// Iteration count
-    virtual auto count() const -> uint = 0;
+    virtual auto count() const -> size_t = 0;
 
     /// Display information
     virtual void disp() const = 0;
 
     //-------------------------------------------------------------------------
-
   };
 
   //---------------------------------------------------------------------------
 
 private:
-
   Interval const T_;
-  int const sign_;
+  int const      sign_;
 
   //
   real t_;
-
 };
 
 } /* namespace dolfin */
