+++ include/dolfin/fem/DofMapCache.h
@@ -18,6 +18,8 @@ class DofMap;
 class Form;
 class Function;
 
+//-----------------------------------------------------------------------------
+
 class DofMapCache
 {
 
@@ -25,12 +27,12 @@ class DofMapCache
 
   struct token_t
   {
-    uint count;
+    size_t         count;
     DofMap * const item;
 
-    token_t(DofMap * dm) :
-        count(1),
-        item(dm)
+    token_t( DofMap * dm )
+      : count( 1 )
+      , item( dm )
     {
     }
   };
@@ -48,30 +50,27 @@ class DofMapCache
 #endif
 
 public:
-
   ///
   void disp() const;
 
 protected:
-
   /// Meyers singleton
-  static auto instance() -> DofMapCache&
+  static auto instance() -> DofMapCache &
   {
     static DofMapCache instance_;
     return instance_;
   }
 
   /// Return a dofmap corresponding to the i-th coefficient space
-  auto acquire(Mesh& mesh, Form const& form, uint const& i) -> DofMap&;
+  auto acquire( Mesh & mesh, Form const & form, size_t const & i ) -> DofMap &;
 
   /// Return a dofmap for the UFC dofmap object
-  auto acquire(Mesh& mesh, ufc::dofmap& dofmap, bool owner) -> DofMap&;
+  auto acquire( Mesh & mesh, ufc::dofmap & dofmap, bool owner ) -> DofMap &;
 
   ///
-  void release(DofMap& dofmap);
+  void release( DofMap & dofmap );
 
 private:
-
   ///
   DofMapCache();
 
@@ -79,11 +78,11 @@ private:
   ~DofMapCache();
 
   container_t cache_;
-  rlist_t rlist_;
-
+  rlist_t     rlist_;
 };
 
-}
+//-----------------------------------------------------------------------------
+
+} // namespace dolfin
 
 #endif /* __DOLFIN_DOF_MAP_CACHE_H */
-
